[hotfix][runtime] Make CheckpointStatsTracker not rely on numberOfTotalSubtasks
I double-checked that the JobInitializationMetricsBuilder constructor is called for every local restart of tasks and the JobInitializationMetricsBuilder#reportInitializationMetrics is called by the deployed Subtask after restoring its state. ExecutionAttemptID is correct because the builder creation happens again when a new Execution attempt is triggered (through the local restart). The totalNumberOfSubtasks was a bit misleading here, in my opinion.